Update attributes in xml and in code

+++ examples/tv-app/linux/include/target-navigator/TargetNavigatorManager.h
@@ -25,5 +25,5 @@ public:
     std::list<chip::app::Clusters::TargetNavigator::Structs::TargetInfo::Type> HandleGetTargetList() override;
     uint8_t HandleGetCurrentTarget() override;
     chip::app::Clusters::TargetNavigator::Commands::NavigateTargetResponse::Type
-    HandleNavigateTarget(const uint64_t & target, const chip::CharSpan & data) override;
+    HandleNavigateTarget(const uint8_t & target, const chip::CharSpan & data) override;
 };
